Handle preferences upgrade from TkRat 2.1.5 to 2.2.
TkRat 2.1.5 was released well after development started on TkRat 2.2
and has a last_version_date that is much later than many of the
development versions of 2.2, but does not contain their preferences
schema changes.  Treat the 2.1.5 last_version_date specially so that
those schema changes are applied when upgrading a 2.1.5 preferences
(~/.ratatosk/ratatoskrc) file.