Fix writing of zero length strings by RatWriteString().
fwrite(ptr, size, n, fp) returns n on a successful write except if
size is zero, in which case the return value is zero.

RatWriteString() passes the string length in the size parameter and
sets n to 1, expecting a return value of 1 to indicate a succesful
write.  This fails when the string has a length of zero, causing
RatWriteString() to return a failure status.

Fix this by swapping the size and n parameters and checking for
the fwrite() return value to be the string length.

This fixes a send failure in SMTP mode with an empty password.