Don't assign an arbitrary value to an enum variable.
Assigning an arbitrary value to an enum variable can produce undefined
behavior.  In this case a -1 value was assigned and the compiler treated
the subsequent == -1 test as always false because it treats enums values
as unsigned.

Restructure the code a bit to avoid the need to set the enum to an
out of range value to detect the none-of-the-above case.