Work around issue of compilers treating enums as unsigned.
The ratDbase code is assigning negative values to enum variables and
then then checking for those values later.  Assiging arbitrary values
to enum variables can result in undefined behavior.  In this case the
compiler is treating the enum variable as unsigned, so the result of
the test will always be false.  As a temporary workaround, cast the
enum to int before the comparision.